remove selected_idx && fix overflow

+++ src/app.rs
@@ -1,14 +1,11 @@
-use open;
 use clipboard;
-use clipboard::ClipboardProvider;
 use clipboard::ClipboardContext;
-use termion::color::LightBlack;
+use clipboard::ClipboardProvider;
+use open;
 use tui::widgets::ListState;
 
-use crate::models::bookmarks;
 use crate::models::bookmarks::Bookmark;
 use rusqlite::{params, Connection};
-
 
 pub struct App {
     pub current_mode: Mode,
@@ -16,7 +13,6 @@ pub struct App {
     pub new_bookmark_name: String,
     pub bookmarks: Vec<Bookmark>,
     pub filtered_bookmarks: Vec<Bookmark>,
-    pub selected_bookmark_idx: usize,
     pub bookmarks_state: ListState,
 }
 
@@ -28,7 +24,6 @@ impl App {
             new_bookmark_name: String::from(""),
             filtered_bookmarks: bookmarks.clone(),
             bookmarks: bookmarks,
-            selected_bookmark_idx: 0,
             bookmarks_state: ListState::default(),
         }
     }
@@ -42,20 +37,10 @@ impl App {
         *self.select_field() = String::from("");
     }
     pub fn on_up(&mut self) {
-        let i = match self.bookmarks_state.selected() {
-            Some(i) => {
-                if i >= self.filtered_bookmarks.len() - 1 {
-                    0
-                } else {
-                    i + 1
-                }
-            }
-            None => 0,
+        if self.filtered_bookmarks.len() == 0 {
+            return self.bookmarks_state.select(None);
         };
-        self.bookmarks_state.select(Some(i)); 
-    }
 
-    pub fn on_down(&mut self) {
         let i = match self.bookmarks_state.selected() {
             Some(i) => {
                 if i == 0 {
@@ -68,37 +53,68 @@ impl App {
         };
         self.bookmarks_state.select(Some(i));
     }
-    pub fn on_delete(&mut self) {
+
+    pub fn on_down(&mut self) {
         if self.filtered_bookmarks.len() == 0 {
-            return
-        }
+            return self.bookmarks_state.select(None);
+        };
 
-        let id_for_delete = &self.filtered_bookmarks[self.selected_bookmark_idx].id;
-        let conn = Connection::open("fbmark.db").unwrap();
-
-        conn.execute(
-            "DELETE FROM bookmarks WHERE id=?1",
-            params![id_for_delete],
-        ).unwrap();
-
-        self.bookmarks = Bookmark::collect_all().unwrap();
-        self.filtered_bookmarks.remove(self.selected_bookmark_idx);
-
-        if self.selected_bookmark_idx != 0 && self.selected_bookmark_idx >= self.filtered_bookmarks.len() {
-            self.selected_bookmark_idx -= 1;
-        }
+        let i = match self.bookmarks_state.selected() {
+            Some(i) => {
+                if i >= self.filtered_bookmarks.len() - 1 {
+                    0
+                } else {
+                    i + 1
+                }
+            }
+            None => 0,
+        };
+        self.bookmarks_state.select(Some(i));
     }
+
+    pub fn on_delete(&mut self) {
+        if self.filtered_bookmarks.len() == 0 || self.bookmarks_state.selected() == None {
+            return;
+        }
+
+        match self.bookmarks_state.selected() {
+            Some(index) => {
+                let id_for_deletion = self.filtered_bookmarks[index].id.clone();
+                let conn = Connection::open("fbmark.db").unwrap();
+                conn.execute(
+                    "DELETE FROM bookmarks WHERE id=?1",
+                    params![id_for_deletion],
+                )
+                .unwrap();
+
+                self.bookmarks = Bookmark::collect_all().unwrap();
+                self.filtered_bookmarks.remove(index);
+                if self.filtered_bookmarks.len() == 0 {
+                    self.bookmarks_state.select(None)
+                } else {
+                    self.on_down()
+                }
+            }
+            None => panic!("Index not found"),
+        };
+    }
+
     pub fn resolve_enter(&mut self) {
         match self.current_mode {
-            Mode::Search => {
-                let url = self.filtered_bookmarks[self.selected_bookmark_idx].url();
-                open::that(url).unwrap();
-            }
+            Mode::Search => match self.bookmarks_state.selected() {
+                Some(index) => {
+                    let url = self.filtered_bookmarks[index].url();
+                    open::that(url).unwrap();
+                }
+                None => {
+                    panic!("Index cannot be found")
+                }
+            },
             Mode::AddBookmark => {
                 let bmark_name = self.new_bookmark_name.clone();
                 match Bookmark::create(bmark_name) {
                     Ok(bmark) => self.bookmarks.push(bmark),
-                    Err(e) => panic!(e)
+                    Err(e) => panic!("{}", e),
                 }
                 self.new_bookmark_name = "".to_string();
                 self.current_mode = Mode::Search;
@@ -108,15 +124,15 @@ impl App {
 
     pub fn select_field(&mut self) -> &mut String {
         match self.current_mode {
-            Mode::Search => { &mut self.search_string },
-            Mode::AddBookmark => { &mut self.new_bookmark_name },
+            Mode::Search => &mut self.search_string,
+            Mode::AddBookmark => &mut self.new_bookmark_name,
         }
     }
 
     pub fn change_mode(&mut self) {
         match self.current_mode {
-            Mode::Search => { self.current_mode = Mode::AddBookmark },
-            Mode::AddBookmark => { self.current_mode = Mode::Search },
+            Mode::Search => self.current_mode = Mode::AddBookmark,
+            Mode::AddBookmark => self.current_mode = Mode::Search,
         }
     }
 
@@ -127,17 +143,17 @@ impl App {
                 for c in payload.chars() {
                     self.select_field().push(c);
                 }
-            },
+            }
             Err(e) => {
                 // panic!(e)
-            },
+            }
         };
     }
 }
 
 pub enum Mode {
     Search,
-    AddBookmark
+    AddBookmark,
 }
 
 pub enum Event<I> {
